+++ Assets/Scripts/SS3D/Systems/Tile/TileMapCreator/TileMapCreator.cs
@@ -38,20 +38,14 @@ namespace SS3D.Systems.Tile.TileMapCreator
         private TMP_Dropdown _layerPlacementDropdown;
         [SerializeField]
         private TMP_InputField _inputField;
-        [SerializeField]
-        private Material _validConstruction;
-        [SerializeField]
-        private Material _invalidConstruction;
-        [SerializeField]
-        private Material _deleteConstruction;
 
+        private bool _mouseOverUI = false;
+        public bool MouseOverUI => _mouseOverUI;
 
         private bool _enabled = false;
-        private bool _itemPlacement = false;
-        private bool _mouseOverUI = false;
 
-        private Vector3 _lastSnappedPosition;
-        private GenericObjectSo _selectedObject;
+
+
 
         private TileSystem _tileSystem;
 
@@ -59,10 +53,6 @@ namespace SS3D.Systems.Tile.TileMapCreator
         private Controls.TileCreatorActions _controls;
         private InputSystem _inputSystem;
         private PanelTab _tab;
-
-        private Vector3 _dragStartPostion;
-        private bool _isDragging;
-        private bool _isDeleting;
 
         [SerializeField]
         private BuildGhostManager _ghostManager;
@@ -77,102 +67,15 @@ namespace SS3D.Systems.Tile.TileMapCreator
             _controls = _inputSystem.Inputs.TileCreator;
             _inputSystem.ToggleAction(_controls.ToggleMenu, true);
             _controls.ToggleMenu.performed += HandleToggleMenu;
-            _controls.Place.started += HandlePlaceStarted;
-            _controls.Place.performed += HandlePlacePerformed;
-            _controls.Replace.performed += HandleReplace;
-            _controls.Replace.canceled += HandleReplace;
-            _controls.Rotate.performed += HandleRotate;
         }
         
         private void HandleUpdate(ref EventContext context, in UpdateEvent updateEvent)
         {
-            ActivateGhosts();
             AdjustGridWidth();
-            
-            Vector3 position = TileHelper.GetPointedPosition(!_itemPlacement);
-            // Move buildGhost, that sticks to the mouse. Currently it exists only if player is not dragging.
-            if (_ghostManager._ghosts.Count == 1)
-            {
-                _ghostManager._ghosts.First().position = position;
-                if (position != _lastSnappedPosition)
-                {
-                    RefreshGhost(_ghostManager._ghosts.First());
-                }
-            }
-
-            if (_isDragging && (position != _lastSnappedPosition) && (_selectedObject != null))
-            {
-                // Delete all ghosts and instantiate new on correct positions. Currently it causes large fps drops.
-                _ghostManager.DestroyGhosts();
-                if (_controls.SquareDrag.phase == InputActionPhase.Performed)
-                {
-                    SquareDrag(position);
-                }
-                else
-                {
-                    LineDrag(position);
-                }
-            }
-            _lastSnappedPosition = position;
         }
 
         #region Handlers
-        public void HandleDeleteButton()
-        {
-            _isDeleting = true;
-        }
-
-        public void HandleBuildButton()
-        {
-            _isDeleting = false;
-        }
-
-        private void HandleRotate(InputAction.CallbackContext context)
-        {
-            _ghostManager.SetNextRotation();
-        }
-
-        private void HandlePlaceStarted(InputAction.CallbackContext context)
-        {
-            // Dragging is disabled for items
-            if (_itemPlacement)
-                return;
-            
-            _isDragging = true;
-            _dragStartPostion = TileHelper.GetPointedPosition(true);
-        }
-
-        private void HandlePlacePerformed(InputAction.CallbackContext context)
-        {
-            _isDragging = false;
-            
-            if (_mouseOverUI)
-            {
-                _inputSystem.ToggleAction(_controls.Place, false);
-            }
-            
-            if (!_isDeleting)
-            {
-                PlaceOnGhosts();
-            }
-            else
-            {
-                DeleteOnGhosts();
-            }
-
-            _ghostManager.DestroyGhosts();
-            
-            if (_selectedObject == null) return;
-            _ghostManager.CreateGhost(_selectedObject.prefab, TileHelper.GetPointedPosition(!_itemPlacement));
-        }
-
-        private void HandleReplace(InputAction.CallbackContext context)
-        {
-            foreach (BuildGhost buildGhost in _ghostManager._ghosts)
-            {
-                RefreshGhost(buildGhost);
-            }
-        }
+       
         
         private void HandleToggleMenu(InputAction.CallbackContext context)
         {
@@ -192,167 +95,10 @@ namespace SS3D.Systems.Tile.TileMapCreator
             LoadObjectGrid(new[] { TileLayer.Plenum }, false);
         }
         
-        /// <summary>
-        /// Place all objects form buildGhost
-        /// </summary>
-        private void PlaceOnGhosts()
-        {
-            bool isReplacing = _controls.Replace.phase == InputActionPhase.Performed;
-
-            foreach (BuildGhost buildGhost in _ghostManager._ghosts)
-            {
-                _tileSystem.RpcPlaceObject(_selectedObject.nameString, buildGhost.position, buildGhost.direction, isReplacing);
-            }
-        }
-
-        /// <summary>
-        /// Delete all objects, that are under ghosts
-        /// </summary>
-        private void DeleteOnGhosts()
-        {
-            if (_itemPlacement)
-            {
-                FindAndDeleteItem();
-            }
-            else
-            {
-                foreach (BuildGhost buildGhost in _ghostManager._ghosts)
-                {
-                    _tileSystem.RpcClearTileObject(_selectedObject.nameString, buildGhost.position, buildGhost.direction);
-                }
-            }
-        }
+       
         #endregion
 
-        /// <summary>
-        /// Update material of buildGhost based build (or anything else) mode and ghosts position  
-        /// </summary>
-        public void RefreshGhost(BuildGhost buildGhost)
-        {
-            if (_isDeleting)
-            {
-                buildGhost.ChangeGhostColor(BuildMatMode.Delete);
-            }
-            else if (_itemPlacement)
-            {
-                buildGhost.ChangeGhostColor(BuildMatMode.Valid);
-            }
-            else
-            {
-                bool isReplacing = _controls.Replace.phase == InputActionPhase.Performed;
-                RpcSendCanBuild(_selectedObject.nameString, buildGhost.position, buildGhost.direction, isReplacing, LocalConnection);
-            }
-        }
 
-
-        [ServerRpc(RequireOwnership = false)]
-        public void RpcSendCanBuild(string tileObjectSoName, Vector3 placePosition, Direction dir, bool replaceExisting, NetworkConnection conn)
-        {
-
-            var tileSystem = Subsystems.Get<TileSystem>();
-
-            TileObjectSo tileObjectSo = (TileObjectSo)tileSystem.GetAsset(tileObjectSoName);
-
-            if (tileObjectSo == null)
-            {
-                Log.Error(this, "Asset is not found");
-                return;
-            }
-
-            bool canBuild = tileSystem.CanBuild(tileObjectSo, placePosition, dir, replaceExisting);
-            RpcReceiveCanBuild(conn, placePosition, canBuild);
-        }
-
-        [TargetRpc]
-        private void RpcReceiveCanBuild(NetworkConnection conn, Vector3 placePosition, bool canBuild)
-        {
-            // Find correct buildGhost to update its material
-            for (int i = 0; i < _ghostManager._ghosts.Count; i++)
-            {
-                BuildGhost buildGhost = _ghostManager._ghosts[i];
-                if (buildGhost.position != placePosition) continue;
-
-                if (canBuild)
-                {
-                    buildGhost.ChangeGhostColor(BuildMatMode.Valid);
-                }
-                else
-                {
-                    buildGhost.ChangeGhostColor(BuildMatMode.Invalid);
-                }
-                return;
-            }
-        }
-        
-        /// <summary>
-        /// Activate all buildGhosts. This method is important, because for some reason network objects disable themselves after a few frames.
-        /// </summary>
-        private void ActivateGhosts()
-        {
-            foreach (BuildGhost buildGhost in _ghostManager._ghosts)
-            {
-                if (!buildGhost.ghost.activeSelf)
-                {
-                    buildGhost.ghost.SetActive(true);
-                }
-            }
-        }
-        
-
-
-        private void FindAndDeleteItem()
-        {
-            Ray ray = Camera.main.ScreenPointToRay(Input.mousePosition);
-            if (Physics.Raycast(ray, out RaycastHit hitInfo))
-            {
-                PlacedItemObject placedItem = hitInfo.collider.gameObject.GetComponent<PlacedItemObject>();
-                if (placedItem != null)
-                {
-                    _tileSystem.RpcClearItemObject(placedItem.NameString, placedItem.gameObject.transform.position);
-                }
-            }
-        }
-
-
-        #region Drag algorithms
-        private void LineDrag(Vector3 position)
-        {
-            Vector2 firstPoint = new(_dragStartPostion.x, _dragStartPostion.z);
-            Vector2 secondPoint = new(position.x, position.z);
-            Vector3[] tiles = MathUtility.FindTilesOnLine(firstPoint, secondPoint)
-                .Select(x => new Vector3(x.x, position.y, x.y)).ToArray();
-            
-            foreach (Vector3 tile in tiles)
-            {
-                var ghost = _ghostManager.CreateGhost(_selectedObject.prefab, tile);
-                RefreshGhost(ghost);
-            }
-        }
-
-        private void SquareDrag(Vector3 position)
-        {
-            int x1 = (int)Math.Min(_dragStartPostion.x, position.x);
-            int x2 = (int)Math.Max(_dragStartPostion.x, position.x);
-            int y1 = (int)Math.Min(_dragStartPostion.z, position.z);
-            int y2 = (int)Math.Max(_dragStartPostion.z, position.z);
-
-            for (int i = y1; i <= y2; i++)
-            {
-                for (int j = x1; j <= x2; j++)
-                {
-                    Vector3 tile = new(j, 0, i);
-                    BuildGhost buildGhost = _ghostManager.CreateGhost(_selectedObject.prefab, tile);
-                    RefreshGhost(buildGhost);
-                }
-            }
-        }
-        
-        #endregion
-        
-        #region Ghosts 
-     
-        
-        #endregion
 
         #region UI
         private void ShowUI(bool show)
@@ -370,7 +116,7 @@ namespace SS3D.Systems.Tile.TileMapCreator
         {
             _mouseOverUI = true;
             _inputSystem.ToggleBinding("<Mouse>/scroll/y", false);
-            if (!_isDragging)
+            if (!_ghostManager.IsDragging)
             {
                 _inputSystem.ToggleAction(_controls.Place, false);
             }
@@ -380,7 +126,7 @@ namespace SS3D.Systems.Tile.TileMapCreator
         {
             _mouseOverUI = false;
             _inputSystem.ToggleBinding("<Mouse>/scroll/y", true);
-            if (!_isDragging)
+            if (!_ghostManager.IsDragging)
             {
                 _inputSystem.ToggleAction(_controls.Place, true);
             }
@@ -445,14 +191,14 @@ namespace SS3D.Systems.Tile.TileMapCreator
                         TileLayer.Plenum
                     }, false);
                     break;
-                
+
                 case 1:
                     LoadObjectGrid(new[]
                     {
                         TileLayer.Turf
                     }, false);
                     break;
-                
+
                 case 2:
                     LoadObjectGrid(new[]
                     {
@@ -460,7 +206,7 @@ namespace SS3D.Systems.Tile.TileMapCreator
                         TileLayer.FurnitureTop
                     }, false);
                     break;
-                
+
                 case 3:
                     LoadObjectGrid(new[]
                     {
@@ -468,7 +214,7 @@ namespace SS3D.Systems.Tile.TileMapCreator
                         TileLayer.WallMountLow
                     }, false);
                     break;
-                
+
                 case 4:
                     LoadObjectGrid(new[]
                     {
@@ -480,32 +226,19 @@ namespace SS3D.Systems.Tile.TileMapCreator
                         TileLayer.PipeMiddle
                     }, false);
                     break;
-                
+
                 case 5:
                     LoadObjectGrid(new[] { TileLayer.Overlays }, false);
                     break;
-                
+
                 case 6:
                     LoadObjectGrid(null, true);
                     break;
-                
+
                 default:
                     ClearGrid();
                     break;
             }
-        }
-        public void SetSelectedObject(GenericObjectSo genericObjectSo)
-        {
-            _itemPlacement = genericObjectSo switch
-            {
-                TileObjectSo => false,
-                ItemObjectSo => true,
-                _ => _itemPlacement,
-            };
-            _selectedObject = genericObjectSo;
-
-            _ghostManager.DestroyGhosts();
-            _ghostManager.CreateGhost(genericObjectSo.prefab, TileHelper.GetPointedPosition(!_itemPlacement));
         }
         
         /// <summary>
